fix: Remove counter property to fix ticket selection issue across multiple statuses
The problem here lies in the 'counter' property. Every time a status is deployed, its value changes to 1. Every time we click on a ticket from the LAST deployed status, it resets to zero. However, if we click on a ticket from another status, it takes the value 1. If 'counter' is not an even number, the ticket cannot be selected. The solution is therefore to eliminate 'counter'.